+++ robocorp-python-ls-core/src/robocorp_ls_core/libs/robocop_lib/robocop/exceptions.py
@@ -1,3 +1,6 @@
+import robot.errors
+
+
 class RobocopFatalError(ValueError):
     pass
 
@@ -63,3 +66,39 @@ class RuleParamFailedInitError(RobocopFatalError):
 class RuleReportsNotFoundError(RobocopFatalError):
     def __init__(self, rule, checker):
         super().__init__(f"{checker.__class__.__name__} checker `reports` attribute contains unknown rule `{rule}`")
+
+
+class InvalidReportName(ConfigGeneralError):
+    def __init__(self, report, reports):
+        from robocop.utils import RecommendationFinder
+
+        report_names = sorted(list(reports.keys()) + ["all"])
+        similar = RecommendationFinder().find_similar(report, report_names)
+        msg = f"Provided report '{report}' does not exist. {similar}"
+        super().__init__(msg)
+
+
+class RobotFrameworkParsingError(Exception):
+    def __init__(self):
+        msg = (
+            "Fatal exception occurred when using Robot Framework parsing module. "
+            "Consider updating Robot Framework to recent stable version."
+        )
+        super().__init__(msg)
+
+
+def handle_robot_errors(func):
+    """
+    If the user uses older version of Robot Framework, it many fail while parsing the
+    source code due to bug that is already fixed in the more recent version.
+    """
+
+    def wrap_errors(*args, **kwargs):
+        try:
+            return func(*args, **kwargs)
+        except robot.errors.DataError:
+            raise
+        except:  # noqa
+            raise RobotFrameworkParsingError
+
+    return wrap_errors