+++ robocorp-python-ls-core/src/robocorp_ls_core/libs/robocop_lib/robocop/run.py
@@ -1,7 +1,6 @@
 """
 Main class of Robocop module. Gathers files to be scanned, checkers, parses CLI arguments and scans files.
 """
-import inspect
 import os
 import sys
 from collections import Counter
@@ -15,6 +14,7 @@ from robocop.config import Config
 from robocop.files import get_files
 from robocop.rules import Message
 from robocop.utils import DisablersFinder, FileType, FileTypeChecker, RecommendationFinder, is_suite_templated
+from robocop.utils.file_types import check_model_type, get_resource_with_lang
 
 
 class Robocop:
@@ -44,12 +44,11 @@ class Robocop:
         self.checkers = []
         self.rules = {}
         self.reports = {}
-        self.disabler = None
         self.root = os.getcwd()
         self.config = Config(from_cli=from_cli) if config is None else config
         self.from_cli = from_cli
         if not from_cli:
-            self.config.reports.add("json_report")
+            self.config.reports.append("json_report")
         self.out = self.set_output()
 
     def set_output(self):
@@ -99,8 +98,9 @@ class Robocop:
                 file_type = FileType.GENERAL
             file_type_checker.source = file
             try:
-                model = file_type.get_parser()(str(file))
-                file_type_checker.visit(model)
+                resource_parser = file_type.get_parser()
+                model = get_resource_with_lang(resource_parser, str(file), self.config.language)
+                check_model_type(file_type_checker, model)
                 self.files[file] = (file_type, model)
             except DataError:
                 print(f"Failed to decode {file}. Default supported encoding by Robot Framework is UTF-8. Skipping file")
@@ -109,7 +109,7 @@ class Robocop:
             if resource in self.files and self.files[resource][0].value != FileType.RESOURCE:
                 self.files[resource] = (
                     FileType.RESOURCE,
-                    get_resource_model(str(resource)),
+                    get_resource_with_lang(get_resource_model, str(resource), self.config.language),
                 )
 
     def run_checks(self):
@@ -125,10 +125,10 @@ class Robocop:
             self.reports["file_stats"].files_count = len(self.files)
 
     def run_check(self, ast_model, filename, source=None):
-        found_issues = []
-        self.register_disablers(filename, source)
-        if self.disabler.file_disabled:
+        disablers = DisablersFinder(filename=filename, source=source)
+        if disablers.file_disabled:
             return []
+        found_issues = []
         templated = is_suite_templated(ast_model)
         for checker in self.checkers:
             if checker.disabled:
@@ -136,13 +136,9 @@ class Robocop:
             found_issues += [
                 issue
                 for issue in checker.scan_file(ast_model, filename, source, templated)
-                if not self.disabler.is_rule_disabled(issue)
+                if not disablers.is_rule_disabled(issue)
             ]
         return found_issues
-
-    def register_disablers(self, filename, source):
-        """Parse content of file to find any disabler statements like # robocop: disable=rulename"""
-        self.disabler = DisablersFinder(filename=filename, source=source)
 
     def report(self, rule_msg: Message):
         for report in self.reports.values():
@@ -205,20 +201,9 @@ class Robocop:
         sys.exit()
 
     def load_reports(self):
-        self.reports = {}
-        classes = inspect.getmembers(reports, inspect.isclass)
-        available_reports = "Available reports:\n"
-        for report_class in classes:
-            if not issubclass(report_class[1], reports.Report):
-                continue
-            report = report_class[1]()
-            if not hasattr(report, "name"):
-                continue
-            if "all" in self.config.reports or report.name in self.config.reports:
-                self.reports[report.name] = report
-            available_reports += f"{report.name:20} - {report.description}\n"
+        self.reports = reports.get_reports(self.config.reports)
         if self.config.list_reports:
-            available_reports += "all" + " " * 18 + "- Turns on all available reports"
+            available_reports = reports.list_reports(self.reports)
             print(available_reports)
             sys.exit()
 
@@ -236,7 +221,10 @@ class Robocop:
 
     def make_reports(self):
         for report in self.reports.values():
-            output = report.get_report()
+            if report.name == "sarif":
+                output = report.get_report(self.config, self.rules)
+            else:
+                output = report.get_report()
             if output is not None:
                 self.write_line(output)
 
@@ -269,6 +257,8 @@ def run_robocop():
     try:
         linter = Robocop(from_cli=True)
         linter.run()
+    except robocop.exceptions.RobotFrameworkParsingError:
+        raise
     except robocop.exceptions.RobocopFatalError as err:
         print(f"Error: {err}")
         sys.exit(1)