+++ robocorp-python-ls-core/src/robocorp_ls_core/libs/robocop_lib/robocop/utils/__init__.py
@@ -8,6 +8,8 @@ from robocop.utils.misc import (
     AssignmentTypeDetector,
     RecommendationFinder,
     find_robot_vars,
+    get_errors,
+    get_section_name,
     is_suite_templated,
     issues_to_lsp_diagnostic,
     keyword_col,
@@ -20,7 +22,6 @@ from robocop.utils.misc import (
     parse_assignment_sign_type,
     pattern_type,
     remove_robot_vars,
+    str2bool,
     token_col,
-    get_section_name,
-    get_errors,
 )